refactor(ai-prompt-template): migrate ai-prompt-template to new framework
