+++ spec/03-plugins/43-ai-prompt-template/02-integration_spec.lua
@@ -5,7 +5,7 @@ local PLUGIN_NAME = "ai-prompt-template"
 
 
 
-for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
+for _, strategy in helpers.all_strategies() do
   describe(PLUGIN_NAME .. ": (access) [#" .. strategy .. "]", function()
     local client
 
@@ -426,4 +426,4 @@ for _, strategy in helpers.all_strategies() do if strategy ~= "cassandra" then
 
   end)
 
-end end
+end
